- added minimal USB support (patch from Ben Lunt <fys@cybertrails.com>, updated   and improved by Christophe Bothamy and Volker Ruppert)
Comment from the author:

Attached is a "patch" file detailing what you need to do
to add USB support (UHCI only for now) to your existing
Bochs (2.0.xx) source code.

I use Win32 and VC++ but the source and modifications
should be platform and compiler independant.

Please let me know if this patch some how breaks the
build process of your compilation (Bochs 2.0.0 or above
only).

I would also like any feedback on how this code works
(or doesn't work) on your platform and within your
images. If you explain in as much detail as you can on
how it did or did not work, I will try to add to the source
to help it work on all platforms and images.


git-svn-id: https://svn.code.sf.net/p/bochs/code/trunk@4608 6c8e4198-a580-4521-a02f-c9ae86df1db9